move MonitorTaskJobControllerUtility to the test project so that we can reuse it fix Its fix more PMD issues
